Reorder generic parameters and use default values
The generic parameters are reordered to `<Data, Name, Cause>`.
With the use of defaults, it allows to use `Incident<Data>`
(which is more useful than `Incident<Name>` or `Incident<Cause>`).
For the moment the cause is kept as the last parameter, it may
change depending on usage in a future major version.